Use new lzo project

+++ src/utils/xrCompress/xrCompress.cpp
@@ -145,7 +145,7 @@ xrCompressor::ALIAS* xrCompressor::testALIAS(IReader* base, u32 crc, u32& a_test
 	return NULL;
 }
 
-void xrCompressor::write_file_header(LPCSTR file_name, const u32 &crc, const u32 &ptr, const u32 &size_real, const u32 &size_compressed)
+void xrCompressor::write_file_header(LPCSTR file_name, const u32 &crc, const u32 &ptr, const u32 &size_real, const size_t &size_compressed)
 {
 	u32					file_name_size = (xr_strlen(file_name) + 0)*sizeof(char);
 	u32					buffer_size = file_name_size + 4*sizeof(u32);
@@ -209,7 +209,7 @@ void xrCompressor::CompressOne(LPCSTR path)
 	u32			c_crc32				=	crc32		(src->pointer(),src->length());
 	u32			c_ptr				=	0;
 	u32			c_size_real			=	0;
-	u32			c_size_compressed	=	0;
+	size_t		c_size_compressed	=	0;
 	u32			a_tests				=	0;
 
 	ALIAS*		A					=	testALIAS	(src,c_crc32,a_tests);
@@ -274,12 +274,12 @@ void xrCompressor::CompressOne(LPCSTR path)
 					if (!bFast)
 					{
 						u8*		c_out	= xr_alloc<u8>	(c_size_real);
-						u32		c_orig	= c_size_real;
+						size_t c_orig	= c_size_real;
 						R_ASSERT		(LZO_E_OK	== lzo1x_optimize	(c_data,c_size_compressed,c_out,&c_orig, NULL));
 						R_ASSERT		(c_orig		== c_size_real		);
 						xr_free			(c_out);
 					}//bFast
-					fs_pack_writer->w	(c_data,c_size_compressed);
+					fs_pack_writer->w	(c_data,(u32)c_size_compressed);
 					printf				("%3.1f%%",	100.f*float(c_size_compressed)/float(src->length()));
 					Msg					("%-80s   - OK (%3.1f%%)",path,100.f*float(c_size_compressed)/float(src->length()));
 				}
